+++ src/SFA.DAS.Roatp.Application.UnitTests/Courses/GetProviderDetailsForCourse/GetProviderDetailsForCourseQueryValidatorTests.cs
@@ -29,7 +29,7 @@ namespace SFA.DAS.Roatp.Application.UnitTests.Courses.GetProviderDetailsForCours
         [TestCase(1, 10000001, null, null, true)]
         [TestCase(1, 10000000, null, null, false)]
         [TestCase(1, 100000000, null, null, false)]
-        public async Task Validate_Ukprn(int larsCode, int ukprn, double? lat, double? lon,
+        public async Task Validate_Ukprn(int larsCode, int ukprn, decimal? lat, decimal? lon,
             bool isValid)
         {
             var validator =
@@ -51,8 +51,8 @@ namespace SFA.DAS.Roatp.Application.UnitTests.Courses.GetProviderDetailsForCours
         public async Task Validate_LarsCode(int larsCode, bool isValid)
         {
             var ukprn = 10000001;
-            Double? lat = null;
-            Double? lon = null;
+            decimal? lat = null;
+            decimal? lon = null;
             var validator = new GetProviderDetailsForCourseQueryValidator(_providersReadRepo.Object, _providerCoursesReadRepo.Object);
 
             var result = await validator.TestValidateAsync(new GetProviderDetailsForCourseQuery(larsCode, ukprn, lat, lon));
@@ -69,7 +69,7 @@ namespace SFA.DAS.Roatp.Application.UnitTests.Courses.GetProviderDetailsForCours
         [TestCase(56, 1.75, false, LatLongValidator.LongitudeOutsideUk)]
         [TestCase(56, null, false, LatLongValidator.LatitudeAndNotLongitude)]
         [TestCase(null, 0, false, LatLongValidator.NotLatitudeAndLongitude)]
-        public async Task Validate_LatitudeLongitude(double? lat, double? lon, bool isValid, string ErrorMessage)
+        public async Task Validate_LatitudeLongitude(decimal? lat, decimal? lon, bool isValid, string ErrorMessage)
         {
             var larsCode = 1;
             var ukprn = 10000001;
